Updated api routes permissions

+++ routes/api/user.php
@@ -8,12 +8,21 @@
 */
 Route::group(["middleware" => "auth:api"], function () {
     Route::group(["middleware" => ["role:admin"]], function () {
-        Route::get("users/{id}", "API\UserController@show");
-        Route::get("users", "API\UserController@all");
-        Route::delete("/user/{id}", "API\UserController@destroy");
-        Route::post("/users", "API\UserController@store");
-        Route::put("users/{id}", "API\UserController@update");
-        Route::patch("/users/{userId}/role/{roleId}", "API\UsersRoleController@store");
-        Route::delete("/users/{userId}/role/{roleId}", "API\UsersRoleController@destroy");
+        Route::get("users/{id}", "API\UserController@show")->middleware("permission:read-users");
+
+        Route::get("users", "API\UserController@all")->middleware("permission:read-users");
+
+        Route::delete("/user/{id}", "API\UserController@destroy")->middleware("permission:delete-users");
+
+        Route::post("/users", "API\UserController@store")->middleware("permission:create-users");
+
+        Route::put("users/{id}", "API\UserController@update")->middleware("permission:update-users");
+
+        Route::patch("/users/{userId}/role/{roleId}", "API\UsersRoleController@store")->middleware(
+            "permission:update-users",
+        );
+        Route::delete("/users/{userId}/role/{roleId}", "API\UsersRoleController@destroy")->middleware(
+            "permission:update-users",
+        );
     });
 });
